% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{Response-class}
\alias{Response}
\alias{Response-class}
\title{CLASS Response}
\description{
Represents a HTTP response.
}
\details{
Represents a HTTP response.
}
\section{Fields}{
  \describe{ \item{\code{response}:}{Object of class
  \code{"jobjRef"}} }
}

\section{Methods}{
  \describe{ \item{\code{new(...)}:}{ Create a new
  \code{Response} object. ... is used to define the
  appropriate slots.} \item{\code{body()}:}{Get the body of
  the response as a plain string.}
  \item{\code{bodyAsBytes()}:}{Get the body of the response
  as an array of bytes.} \item{\code{charset()}:}{Get the
  character set name of the response.}
  \item{\code{contentType()}:}{Get the response content
  type (e.g. "text/html")} \item{\code{parse()}:}{Parse the
  body of the response as a \code{\link{Document}}. }
  \item{\code{statusCode()}:}{Get the status code of the
  response.} \item{\code{statusMessage()}:}{Get the status
  code of the response.} }
}

