% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{Connection-class}
\alias{Connection}
\alias{Connection-class}
\title{CLASS Connection}
\description{
A Connection provides a convenient interface to fetch content from the web, and parse them into Documents.
}
\details{
To get a new Connection, use Jsoup$connect(String). Connections contain Connection.Request and Connection.Response objects. The request objects are reusable as prototype requests.
Request configuration can be made using either the shortcut methods in Connection (e.g. userAgent(String)), or by methods in the Connection.Request object directly. All request configuration must be made before the request is executed.
}
\section{Fields}{
  \describe{ \item{\code{connection}:}{Object of class
  \code{"jobjRef"}} }
}

\section{Methods}{
  \describe{ \item{\code{new(...)}:}{ Create a new
  \code{Connection} object. ... is used to define the
  appropriate slots.} \item{\code{cookie(name,
  value)}:}{Set a cookie to be sent in the request.
  \describe{ \item{\code{name: }}{Character: name of cookie
  } \item{\code{value: }}{Character: value of cookie} } }
  \item{\code{cookies(cookies)}:}{Adds each of the supplied
  cookies to the request. \describe{ \item{\code{cookies:
  }}{list: A list of cookie names and value pairs. The list
  values should be of type character.} } }
  \item{\code{get()}:}{Execute the request as a GET, and
  parse the result. Returns a \code{\link{Document}}}
  \item{\code{post()}:}{Execute the request as a POST, and
  parse the result. Returns a \code{\link{Document}}} }
}

